+++ packages/template-retail-react-app/CHANGELOG.md
@@ -1,4 +1,4 @@
-## v8.3.0-dev (Nov 05, 2025)
+## v8.3.0-preview.0 (Dec 12, 2025)
 - [Bugfix] Fix Forgot Password link not working from Account Profile password update form [#3493](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3493)
 - Introduce Address Autocompletion feature in the checkout flow, powered by Google Maps Platform [#3071](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3071)